Moved all client routes into client router, adjusted routes to match documentation
