+++ app/index.js
@@ -14,10 +14,9 @@ const express = require('express');
 const app = express();
 const PORT = process.env.PORT || 8000;
 const path = require('path');
-const flash = require('connect-flash');
-const countryUtils = require('./utils/countryUtils');
 
 const authRouter = require('./routes/auth');
+const clientRouter = require('./routes/client');
 const apiRouter = require('./routes/api');
 const passport = require('passport');
 const session = require('express-session');
@@ -67,95 +66,12 @@ const swaggerDoc = yaml.parse(swaggerFile);
 app.use('/api-docs', swaggerUI.serve, swaggerUI.setup(swaggerDoc));
 
 // ================ ROUTERS ========================
-app.use('/', authRouter);
+app.use('/', authRouter, clientRouter);
 app.use('/api', apiRouter);
 
 // ================ JS AND CSS PATH SETUP ================
 app.use(express.static(path.join(__dirname, 'public/css')));
 app.use(express.static(path.join(__dirname, 'public/js')));
-
-// ================ APP ROUTES ================
-app.get('/', (_, res) => {
-    const pageContext = {
-        // Fake Fact data to mock landing page fact of the day, may be replaced with random fact?
-        factoid: {
-            id: 123,
-            content: 'A super cool 37 fact',
-            // note: 'Something extra about the fact',
-        },
-    };
-    res.render('pages/landing-page', pageContext);
-});
-
-// * This route is used for example EJS usage demonstration. Should be removed after we are familiar with EJS.
-app.get('/example', (_, res) => {
-    // Demonstrating passing an obj with values to a rendered template
-    const pageContext = { injectedVal: 'Superb!' };
-    res.render('pages/example', pageContext);
-});
-
-// This route is for the admin dashboard where the admin can approved, edit, and delete fact submissions.
-app.get('/admin', (req, res) => {
-    const testData = [
-        {
-            dateSubmitted: '05 / 04 / 2024',
-            user: 'abc0185',
-            fact: 'The number of pages in this book is a multiple of 37!',
-            note: 'Name of the book: "All things 37" by Greg Jones',
-            tags: 'media, books',
-        },
-        {
-            dateSubmitted: '05 / 06 / 2024',
-            user: 'mag3737',
-            fact: 'Another super cool 37 fact',
-            note: 'Found in Vancouver, BC',
-            tags: 'nature',
-        },
-    ];
-    const adminName = 'mag3737';
-    res.render('pages/admin-dashboard', { submissions: testData, adminName });
-});
-
-// route for submitting facts
-app.get('/fact_submission', (req, res) => {
-    if (!req.user) {
-        // gets country data from json for fact submitter country options
-        countryUtils.readCountryData((err, countries) => {
-            if (err) {
-                return res.status(500).send('Internal Server Error');
-            }
-            res.render('pages/fact-submission-page', { countries: countries});
-        });
-    } else {
-        res.render('pages/fact-submission-page', { user: req.user });
-    }
-});
-
-// This route is for the factoids listings page where users can view and search for factoids.
-app.get('/factoids', async (req, res) => {
-    const pageContext = {
-        // Fake Fact data to mock factoid, may be replaced with actual data from the database
-        factoid: {
-            id: 777,
-            content: 'A super cool 37 fact',
-            note: 'Something extra about the fact',
-        },
-    };
-    res.render('pages/factoid-listings', pageContext);
-});
-
-// This route is for the about/why 37? page.
-app.get('/about', async (req, res) => {
-    res.render('pages/about');
-});
-
-// This route displays contact information
-app.get('/contact', (req, res) => {
-    res.render('pages/contact');
-});
-
-// ================ SERVER ROUTES ================
-// TODO: Add server REST route calls for making SQLite queries through prisma
 
 // Begin the server and listen on the configured port
 app.listen(PORT);
