+++ app/public/js/register.js
