+++ app/views/pages/login.ejs
@@ -15,7 +15,7 @@
             <input id="current-password" name="password" type="password" class="form-control" autocomplete="current-password" required>
         </div>
         <button type="submit" class="btn btn-primary">Log in</button>
-        <div class="mt-3"><a href="/signup">Don't have an account? Sign up here!</a></div>
+        <div class="mt-3"><a href="/register">Don't have an account? Sign up here!</a></div>
     </form>
 </main>
