+++ app/views/pages/register.ejs
@@ -6,7 +6,7 @@
         <div class="alert alert-danger" role="alert"><%= err %></div>
     <% } %>  
     <h1>Sign Up</h1>
-    <form action="/api/signup" method="post">
+    <form action="/api/register" method="post">
         <div class="mb-3">
             <label for="username" class="form-label">Email</label>
             <input id="email" name="email" type="email" class="form-control" autocomplete="email" required>
@@ -34,7 +34,7 @@
         <div class="mt-3"><a href="/login">Already have an account? Log in here!</a></div>
     </form>
 
-    <script src="signup.js" defer></script>
+    <script src="register.js" defer></script>
 </main>
 
 <%- include('../partials/footer'); %>
